backend/requirements.txt
Flask==3.1.0
paramiko==3.5.0
python-dotenv==1.0.1
waitress==3.0.2
